+++ Telegram/groups.go
@@ -1,4 +1,4 @@
-package wasabeetelegram
+package wtg
 
 import (
 	"bytes"
@@ -6,7 +6,7 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/go-telegram-bot-api/telegram-bot-api"
+	"github.com/go-telegram-bot-api/telegram-bot-api/v5"
 
 	"github.com/wasabee-project/Wasabee-Server/log"
 	"github.com/wasabee-project/Wasabee-Server/messaging"
@@ -21,8 +21,7 @@ func processChatMessage(inMsg *tgbotapi.Update) error {
 }
 
 func processChatCommand(inMsg *tgbotapi.Update) error {
-	tgid := model.TelegramID(inMsg.Message.From.ID)
-	gid, _, err := tgid.GidV()
+	gid, err := model.TelegramID(inMsg.Message.From.ID).Gid()
 	if err != nil {
 		log.Error(err)
 		return err
@@ -38,6 +37,35 @@ func processChatCommand(inMsg *tgbotapi.Update) error {
 	msg.Text = defaultReply
 
 	switch inMsg.Message.Command() {
+	case "unlink":
+		teamID, _, err := model.ChatToTeam(inMsg.Message.Chat.ID)
+		log.Debugw("unlinking team from chat", "chatID", inMsg.Message.Chat.ID, "GID", gid, "resource", teamID)
+
+		owns, err := gid.OwnsTeam(teamID)
+		if err != nil {
+			log.Error(err)
+			return err
+		}
+		if !owns {
+			err = fmt.Errorf("only team owner can unlink the team")
+			log.Error(err)
+			return err
+		}
+
+		if err := teamID.UnlinkFromTelegramChat(model.TelegramID(inMsg.Message.Chat.ID)); err != nil {
+			log.Error(err)
+			msg.Text = err.Error()
+			if _, err := bot.Send(msg); err != nil {
+				log.Error(err)
+			}
+			return err
+		}
+
+		msg.Text = "Successfully unlinked"
+		if _, err := bot.Send(msg); err != nil {
+			log.Error(err)
+			return err
+		}
 	case "link":
 		tokens := strings.Split(inMsg.Message.Text, " ")
 		if len(tokens) > 1 {
@@ -83,18 +111,28 @@ func processChatCommand(inMsg *tgbotapi.Update) error {
 		}
 	case "status":
 		msg.ParseMode = "HTML"
-		teamID, err := model.ChatToTeam(inMsg.Message.Chat.ID)
+		teamID, opID, err := model.ChatToTeam(inMsg.Message.Chat.ID)
 		if err != nil {
-			log.Error(err)
+			// log.Debug(err) // not linked is not an error
 			msg.Text = err.Error()
 			if _, err := bot.Send(msg); err != nil {
 				log.Error(err)
 				return err
 			}
-			return err
+			return nil
 		}
 		name, _ := teamID.Name()
+
 		msg.Text = fmt.Sprintf("Linked to team: <b>%s</b> (%s)", name, teamID.String())
+
+		if opID != "" {
+			opstat, err := opID.Stat()
+			if err != nil {
+				log.Error(err)
+			} else {
+				msg.Text = fmt.Sprintf("%s \nLinked to Operation <b>%s</b> (%s)", msg.Text, opstat.Name, opID)
+			}
+		}
 		if _, err := bot.Send(msg); err != nil {
 			log.Error(err)
 			return err
@@ -116,67 +154,58 @@ func processChatCommand(inMsg *tgbotapi.Update) error {
 		} else {
 			filterGid = ""
 		}
-		teamID, err := model.ChatToTeam(inMsg.Message.Chat.ID)
+		teamID, opID, err := model.ChatToTeam(inMsg.Message.Chat.ID)
 		if err != nil {
 			log.Error(err)
 			if _, err := bot.Send(msg); err != nil {
 				log.Error(err)
-				return err
 			}
 			return err
 		}
-		ops, err := teamID.Operations()
-		if err != nil {
+		if opID == "" {
+			err := fmt.Errorf("team must be linked to operation to view assignments")
+			return err
+		}
+		o := model.Operation{}
+		o.ID = opID
+		if err := o.Populate(gid); err != nil {
 			log.Error(err)
 			return err
 		}
-		for _, p := range ops {
-			var o model.Operation
-			o.ID = p.OpID
-			err := o.Populate(gid)
-			if err != nil {
-				log.Error(err)
+		var b bytes.Buffer
+		name, _ := teamID.Name()
+		b.WriteString(fmt.Sprintf("<b>Operation: %s</b> (team: %s)\n", o.Name, name))
+		b.WriteString("<b>Order / Portal / Action / Agent / State</b>\n")
+		for _, m := range o.Markers {
+			// if the caller requested the results to be filtered...
+			if filterGid != "" && m.AssignedTo != filterGid {
 				continue
 			}
-			var b bytes.Buffer
-			name, _ := teamID.Name()
-			b.WriteString(fmt.Sprintf("<b>Operation: %s</b> (team: %s)\n", o.Name, name))
-			b.WriteString("<b>Order / Portal / Action / Agent / State</b>\n")
-			for _, m := range o.Markers {
-				// if the caller requested the results to be filtered...
-				if filterGid != "" && m.AssignedTo != filterGid {
-					continue
+			if m.State != "pending" && m.AssignedTo != "" {
+				p, _ := o.PortalDetails(m.PortalID, gid)
+				a, _ := m.AssignedTo.IngressName()
+				tg, _ := m.AssignedTo.TelegramName()
+				if tg != "" {
+					a = fmt.Sprintf("@%s", tg)
 				}
-				if m.State != "pending" && m.AssignedTo != "" {
-					p, _ := o.PortalDetails(m.PortalID, gid)
-					a, _ := m.AssignedTo.IngressName()
-					tg, _ := m.AssignedTo.TelegramName()
-					if tg != "" {
-						a = fmt.Sprintf("@%s", tg)
-					}
-					stateIndicatorStart := ""
-					stateIndicatorEnd := ""
-					if m.State == "completed" {
-						stateIndicatorStart = "<strike>"
-						stateIndicatorEnd = "</strike>"
-					}
-					b.WriteString(fmt.Sprintf("%d / %s<a href=\"http://maps.google.com/?q=%s,%s\">%s</a> / %s / %s / %s%s\n",
-						m.Order, stateIndicatorStart, p.Lat, p.Lon, p.Name, model.NewMarkerType(m.Type), a, m.State, stateIndicatorEnd))
+				stateIndicatorStart := ""
+				stateIndicatorEnd := ""
+				if m.State == "completed" {
+					stateIndicatorStart = "<strike>"
+					stateIndicatorEnd = "</strike>"
 				}
+				b.WriteString(fmt.Sprintf("%d / %s<a href=\"http://maps.google.com/?q=%s,%s\">%s</a> / %s / %s / %s%s\n",
+					m.Order, stateIndicatorStart, p.Lat, p.Lon, p.Name, model.NewMarkerType(m.Type), a, m.State, stateIndicatorEnd))
 			}
 			msg.Text = b.String()
 			if _, err := bot.Send(msg); err != nil {
 				log.Error(err)
-				msg.Text = err.Error()
-				if _, err := bot.Send(msg); err != nil {
-					log.Error(err)
-				}
-				continue
 			}
+			continue
 		}
 	case "unassigned":
 		msg.ParseMode = "HTML"
-		teamID, err := model.ChatToTeam(inMsg.Message.Chat.ID)
+		teamID, opID, err := model.ChatToTeam(inMsg.Message.Chat.ID)
 		if err != nil {
 			log.Error(err)
 			if _, err := bot.Send(msg); err != nil {
@@ -185,28 +214,24 @@ func processChatCommand(inMsg *tgbotapi.Update) error {
 			}
 			return err
 		}
-		ops, err := teamID.Operations()
-		if err != nil {
+		if opID == "" {
+			err := fmt.Errorf("team must be linked to operation to view assignments")
+			return err
+		}
+		o := model.Operation{}
+		o.ID = opID
+		if err := o.Populate(gid); err != nil {
 			log.Error(err)
 			return err
 		}
-		for _, p := range ops {
-			var o model.Operation
-			o.ID = p.OpID
-			err := o.Populate(gid)
-			if err != nil {
-				log.Error(err)
-				continue
-			}
-			var b bytes.Buffer
-			name, _ := teamID.Name()
-			b.WriteString(fmt.Sprintf("<b>Operation: %s</b> (team: %s)\n", o.Name, name))
-			b.WriteString("<b>Order / Portal / Action</b>\n")
-			for _, m := range o.Markers {
-				if m.State == "pending" {
-					p, _ := o.PortalDetails(m.PortalID, gid)
-					b.WriteString(fmt.Sprintf("<b>%d</b> / <a href=\"http://maps.google.com/?q=%s,%s\">%s</a> / %s\n", m.Order, p.Lat, p.Lon, p.Name, model.NewMarkerType(m.Type)))
-				}
+		var b bytes.Buffer
+		name, _ := teamID.Name()
+		b.WriteString(fmt.Sprintf("<b>Operation: %s</b> (team: %s)\n", o.Name, name))
+		b.WriteString("<b>Order / Portal / Action</b>\n")
+		for _, m := range o.Markers {
+			if m.State == "pending" {
+				p, _ := o.PortalDetails(m.PortalID, gid)
+				b.WriteString(fmt.Sprintf("<b>%d</b> / <a href=\"http://maps.google.com/?q=%s,%s\">%s</a> / %s\n", m.Order, p.Lat, p.Lon, p.Name, model.NewMarkerType(m.Type)))
 			}
 			msg.Text = b.String()
 			if _, err := bot.Send(msg); err != nil {
@@ -221,26 +246,26 @@ func processChatCommand(inMsg *tgbotapi.Update) error {
 }
 
 func chatResponses(inMsg *tgbotapi.Update) error {
-	// log.Debugw("message in chat", "chatID", inMsg.Message.Chat.ID)
-	teamID, err := model.ChatToTeam(inMsg.Message.Chat.ID)
+	teamID, opID, err := model.ChatToTeam(inMsg.Message.Chat.ID)
 	if err != nil {
 		// no need to log these, just non-linked chats
-		// log.Error(err)
+		log.Debugw("unknown chat", "err", err.Error(), "chatID", inMsg.Message.Chat.ID)
 		return nil
 	}
 
+	log.Debugw("message in chat", "chatID", inMsg.Message.Chat.ID, "team", teamID, "op", opID)
+
+	// if the bot is removed from the chat, unlink the team from the chat
 	if inMsg.Message.LeftChatMember != nil && inMsg.Message.LeftChatMember.ID == bot.Self.ID {
-		if err := teamID.UnlinkFromTelegramChat(inMsg.Message.Chat.ID); err != nil {
+		if err := teamID.UnlinkFromTelegramChat(model.TelegramID(inMsg.Message.Chat.ID)); err != nil {
 			log.Error(err)
 			return err
 		}
 	}
 
+	// when new people are added to the chat, attempt to add them to the team
 	if inMsg.Message.NewChatMembers != nil {
-		// var ncm []tgbotapi.User
-		// ncm = *inMsg.Message.NewChatMembers
-		// for _, new := range ncm {
-		for _, new := range *inMsg.Message.NewChatMembers {
+		for _, new := range inMsg.Message.NewChatMembers {
 			log.Debugw("new chat member", "tgid", new.ID, "tg", new.UserName)
 			tgid := model.TelegramID(new.ID)
 			gid, err := tgid.Gid()
@@ -326,7 +351,11 @@ func AddToChat(g messaging.GoogleID, t messaging.TeamID) error {
 		log.Debug("no chat linked to team")
 		return nil
 	}
-	chat, err := bot.GetChat(tgbotapi.ChatConfig{ChatID: chatID})
+
+	cic := tgbotapi.ChatInfoConfig{}
+	cic.ChatID = chatID
+
+	chat, err := bot.GetChat(cic)
 	if err != nil {
 		log.Errorw(err.Error(), "chatID", chatID, "GID", gid)
 		return err
@@ -371,7 +400,10 @@ func RemoveFromChat(g messaging.GoogleID, t messaging.TeamID) error {
 		return nil
 	}
 
-	chat, err := bot.GetChat(tgbotapi.ChatConfig{ChatID: chatID})
+	cic := tgbotapi.ChatInfoConfig{}
+	cic.ChatID = chatID
+
+	chat, err := bot.GetChat(cic)
 	if err != nil {
 		log.Errorw(err.Error(), "chatID", chatID, "GID", gid)
 		return err
