+++ model/database.go
@@ -70,7 +70,7 @@ func setupTables() {
 		{"portal", `CREATE TABLE portal ( ID varchar(41) NOT NULL, opID char(40) NOT NULL, name varchar(128) NOT NULL, loc point NOT NULL, comment text DEFAULT NULL, hardness varchar(64) DEFAULT NULL, PRIMARY KEY (ID,opID), KEY fk_operation_id (opID)) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;`},
 		{"rocks", `CREATE TABLE rocks ( gid char(21) NOT NULL, tgid int(11) DEFAULT NULL, agent varchar(64) DEFAULT NULL, verified tinyint(4) NOT NULL DEFAULT 0, smurf tinyint(4) NOT NULL DEFAULT 0, fetched datetime NOT NULL DEFAULT current_timestamp(), PRIMARY KEY (gid), CONSTRAINT fk_rocks_gid FOREIGN KEY (gid) REFERENCES agent (gid) ON DELETE CASCADE) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;`},
 		{"telegram", `CREATE TABLE telegram (telegramID bigint(20) NOT NULL, telegramName varchar(32) NOT NULL, gid char(21) NOT NULL, verified tinyint(1) NOT NULL DEFAULT 0, authtoken varchar(32) DEFAULT NULL, PRIMARY KEY (telegramID), UNIQUE KEY gid (gid), CONSTRAINT fk_agent_telegram FOREIGN KEY (gid) REFERENCES agent (gid) ON DELETE CASCADE) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;`},
-		{"telegramteam", `CREATE TABLE telegramteam (teamID varchar(64) NOT NULL, telegram bigint(20) NOT NULL, opID char(21) DEFAULT NULL, PRIMARY KEY (teamID,telegram), KEY (teamID), CONSTRAINT fk_tt_team FOREIGN KEY (teamID) REFERENCES team (teamID) ON DELETE CASCADE, KEY (telegram), KEY (opID), CONSTRAINT fk_tt_op FOREIGN KEY (opID) REFERENCES operation (ID) ON DELETE CASCADE) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;`},
+		{"telegramteam", `CREATE TABLE telegramteam (teamID varchar(64) NOT NULL, telegram bigint(20) NOT NULL, opID char(40) DEFAULT NULL, PRIMARY KEY (teamID,telegram), KEY (teamID), CONSTRAINT fk_tt_team FOREIGN KEY (teamID) REFERENCES team (teamID) ON DELETE CASCADE, UNIQUE KEY (telegram), KEY (opID), CONSTRAINT fk_tt_op FOREIGN KEY (opID) REFERENCES operation (ID) ON DELETE CASCADE) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;`},
 		{"v", `CREATE TABLE v (gid char(21) NOT NULL, enlid char(42) NOT NULL, vlevel int(11) DEFAULT NULL, vpoints int(11) DEFAULT NULL, agent varchar(64) DEFAULT NULL, level int(11) DEFAULT NULL, quarantine tinyint(4) NOT NULL DEFAULT 0, active tinyint(4) NOT NULL DEFAULT 0, blacklisted tinyint(4) NOT NULL DEFAULT 0, verified tinyint(4) NOT NULL DEFAULT 0, flagged tinyint(4) NOT NULL DEFAULT 0, banned tinyint(4) NOT NULL DEFAULT 0, cellid varchar(32) DEFAULT NULL, telegram varchar(32) DEFAULT NULL, telegramID int(11) NOT NULL DEFAULT 0, startlat float DEFAULT NULL, startlon float DEFAULT NULL, distance int(11) DEFAULT NULL, vapikey char(42) DEFAULT NULL, fetched datetime NOT NULL DEFAULT utc_timestamp(), PRIMARY KEY (gid), CONSTRAINT fk_v_gid FOREIGN KEY (gid) REFERENCES agent (gid) ON DELETE CASCADE) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;`},
 		{"zone", `CREATE TABLE zone (ID tinyint(4) NOT NULL, opID char(40) NOT NULL, name varchar(64) NOT NULL DEFAULT 'zone', color varchar(10) NOT NULL DEFAULT 'green', PRIMARY KEY (ID,opID), KEY fk_operation_zone (opID), CONSTRAINT fk_operation_zone FOREIGN KEY (opID) REFERENCES operation (ID) ON DELETE CASCADE) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;`},
 		{"zonepoints", `CREATE TABLE zonepoints ( zoneID tinyint(4) NOT NULL, opID char(40) NOT NULL, position tinyint(4) NOT NULL, point point NOT NULL, PRIMARY KEY (zoneID,opID,position), KEY fk_operation_zonepoint (opID), CONSTRAINT fk_operation_zonepoint FOREIGN KEY (opID) REFERENCES operation (ID) ON DELETE CASCADE) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;`},
