+++ model/team.go
@@ -2,7 +2,6 @@ package model
 
 import (
 	"database/sql"
-	"encoding/json"
 	"fmt"
 	"strconv"
 
@@ -584,106 +583,4 @@ func (teamID TeamID) JoinToken(gid GoogleID, key string) error {
 	}
 
 	return nil
-}
-
-// LinkToTelegramChat associates a telegram chat ID with the team, performs authorization
-func (teamID TeamID) LinkToTelegramChat(chat TelegramID, opID OperationID) error {
-	_, err := db.Exec("REPLACE INTO telegramteam (teamID, telegram, opID) VALUES (?,?,?)", teamID, chat, MakeNullString(opID))
-	if err != nil {
-		log.Error(err)
-		return err
-	}
-
-	log.Infow("linked team to telegram", "resource", teamID, "chatID", chat, "opID", opID)
-	return nil
-}
-
-// UnlinkFromTelegramChat disassociates a telegram chat ID from the team -- not authenticated since bot removal from chat is enough
-func (teamID TeamID) UnlinkFromTelegramChat(chat int64) error {
-	_, err := db.Exec("DELETE FROM telegramteam WHERE teamID = ? AND telegram = ?", teamID, chat)
-	if err != nil {
-		log.Error(err)
-		return err
-	}
-
-	log.Infow("unlinked team from telegram", "resource", teamID, "chatID", chat)
-	return nil
-}
-
-// TelegramChat returns the associated telegram chat ID for this team, if any
-func (teamID TeamID) TelegramChat() (int64, error) {
-	var chatID int64
-
-	err := db.QueryRow("SELECT telegram FROM telegramteam WHERE teamID = ?", teamID).Scan(&chatID)
-	if err != nil && err != sql.ErrNoRows {
-		log.Error(err)
-		return chatID, err
-	}
-	if err == sql.ErrNoRows {
-		return chatID, nil
-	}
-	return chatID, nil
-}
-
-// ChatToTeam takes a chatID and returns a linked teamID
-func ChatToTeam(chat int64) (TeamID, OperationID, error) {
-	var t TeamID
-	var o OperationID
-
-	err := db.QueryRow("SELECT teamID, opID FROM telegramteam WHERE telegram = ?", chat).Scan(&t, &o)
-	if err != nil && err != sql.ErrNoRows {
-		log.Error(err)
-		return t, o, err
-	}
-	if err == sql.ErrNoRows {
-		err := fmt.Errorf("chat not linked to any teams")
-		return t, o, err
-	}
-	return t, o, nil
-}
-
-// GetAgentLocations is a fast-path to get all available agent locations
-func (gid GoogleID) GetAgentLocations() (string, error) {
-	type loc struct {
-		Gid  GoogleID `json:"gid"`
-		Lat  float64  `json:"lat"`
-		Lon  float64  `json:"lng"`
-		Date string   `json:"date"`
-	}
-
-	var list []loc
-	var tmpL loc
-	var lat, lon string
-
-	var rows *sql.Rows
-	rows, err := db.Query("SELECT x.gid, Y(l.loc), X(l.loc), l.upTime "+
-		"FROM agentteams=x, locations=l "+
-		"WHERE x.teamID IN (SELECT teamID FROM agentteams WHERE gid = ?) "+
-		"AND x.shareLoc= 1 AND x.gid = l.gid", gid)
-	if err != nil {
-		log.Error(err)
-		return "", err
-	}
-
-	defer rows.Close()
-	for rows.Next() {
-		if err := rows.Scan(&tmpL.Gid, &lat, &lon, &tmpL.Date); err != nil {
-			log.Error(err)
-			return "", err
-		}
-		tmpL.Lat, _ = strconv.ParseFloat(lat, 64)
-		tmpL.Lon, _ = strconv.ParseFloat(lon, 64)
-
-		if tmpL.Lat == 0 || tmpL.Lon == 0 {
-			continue
-		}
-
-		list = append(list, tmpL)
-	}
-
-	jList, err := json.Marshal(list)
-	if err != nil {
-		return "", err
-	}
-	return string(jList), nil
 }
